requirements.txt
jupyter
numpy
pandas
scikit-learn
torch
torchaudio
tqdm
pyyaml
